feat(notice): refine types & enums
