+++ src/common/enums/cookie.ts
@@ -1,0 +1,2 @@
+export const COOKIE_TOKEN_NAME = '__token'
+export const COOKIE_USER_GROUP = '__user_group'
