+++ src/common/enums/file.ts
@@ -1,0 +1,20 @@
+export const ASSET_TYPE = {
+  avatar: 'avatar',
+  cover: 'cover',
+  embed: 'embed',
+  embedaudio: 'embedaudio',
+  profileCover: 'profileCover',
+  oauthClientAvatar: 'oauthClientAvatar',
+  tagCover: 'tagCover',
+}
+
+export const ACCEPTED_UPLOAD_IMAGE_TYPES: string[] = [
+  'image/gif',
+  'image/png',
+  'image/jpeg',
+  'image/webp',
+]
+
+export const ACCEPTED_UPLOAD_AUDIO_TYPES: string[] = ['audio/mpeg', 'audio/aac']
+
+export const ACCEPTED_UPLOAD_MIGRATION_TYPES: string[] = ['text/html']
