+++ src/common/enums/hardcode.ts
@@ -1,0 +1,26 @@
+export const INVALID_NAMES = [
+  'administrator',
+  'administrators',
+  'admin',
+  'admins',
+  'team',
+  'teams',
+  'matters',
+  'member',
+  'members',
+  'user',
+  'users',
+  'matty',
+  'matties',
+  '團隊',
+  '团队',
+  '管理員',
+  '管理员',
+  'mattersadmin',
+  'matters團隊',
+  'matters团队',
+  'matters管理員',
+  'matters管理员',
+]
+
+export const AUTO_FOLLOW_TAGS = ['Matters新人打卡', '玩轉Matters實用指南']
