+++ src/common/enums/language.ts
@@ -1,0 +1,5 @@
+export const LANGUAGE = {
+  zh_hans: 'zh_hans',
+  zh_hant: 'zh_hant',
+  en: 'en',
+}
