+++ src/common/enums/table.ts
@@ -1,0 +1,26 @@
+export const VIEW = {
+  articleValue: 'article_value_view',
+  articleActivity: 'article_activity_view',
+
+  /* Hottest articles A/B test */
+  articleHottestA: 'article_hottest_a_view',
+  articleHottestB: 'article_hottest_b_view',
+}
+
+export const MATERIALIZED_VIEW = {
+  articleCountMaterialized: 'article_count_materialized',
+  tagCountMaterialized: 'tag_count_materialized',
+  userReaderMaterialized: 'user_reader_materialized',
+  articleActivityMaterialized: 'article_activity_materialized',
+  articleValueMaterialized: 'article_value_materialized',
+  featuredCommentMaterialized: 'featured_comment_materialized',
+  curationTagMaterialized: 'curation_tag_materialized',
+
+  /* Hottest articles A/B test */
+  articleHottestAMaterialized: 'article_hottest_a_materialized',
+  articleHottestBMaterialized: 'article_hottest_b_materialized',
+
+  mostActiveAuthorMaterialized: 'most_active_author_materialized',
+  mostAppreciatedAuthorMaterialized: 'most_appreciated_author_materialized',
+  mostTrendyAuthorMaterialized: 'most_trendy_author_materialized',
+}
