+++ src/common/enums/time.ts
@@ -1,0 +1,9 @@
+export const MINUTE = 1000 * 60
+export const HOUR = MINUTE * 60
+export const DAY = HOUR * 24
+export const SERVER_TIMEOUT = 5 * MINUTE
+
+export const USER_ACCESS_TOKEN_EXPIRES_IN_MS = DAY * 90 // 90 days
+export const OAUTH_AUTHORIZATION_TOKEN_EXPIRES_IN_MS = MINUTE * 10 // 10 mins
+export const OAUTH_ACCESS_TOKEN_EXPIRES_IN_MS = DAY * 30 // 30 days
+export const OAUTH_REFRESH_TOKEN_EXPIRES_IN_MS = DAY * 90 // 90 days
