+++ src/common/enums/upload.ts
@@ -1,0 +1,9 @@
+export const UPLOAD_FILE_COUNT_LIMIT = 50
+
+export const UPLOAD_FILE_SIZE_LIMIT = 100 * 1024 * 1024
+
+export const UPLOAD_IMAGE_SIZE_LIMIT = 5 * 1024 * 1024
+
+export const UPLOAD_AUDIO_SIZE_LIMIT = 100 * 1024 * 1024
+
+export const UPLOAD_MIGRATION_SIZE_LIMIT = 1 * 1024 * 1024
