+++ src/mutations/article/deleteArticlesTags.ts
@@ -2,7 +2,7 @@ import _difference from 'lodash/difference'
 import _some from 'lodash/some'
 import _uniq from 'lodash/uniq'
 
-import { USER_STATE } from 'common/enums'
+import { DB_NOTICE_TYPE, USER_STATE } from 'common/enums'
 import { environment } from 'common/environment'
 import {
   AuthenticationError,
@@ -68,7 +68,7 @@ const resolver: MutationToDeleteArticlesTagsResolver = async (
   deleteIds.forEach(async (articleId: string) => {
     const article = await articleService.baseFindById(articleId)
     await notificationService.trigger({
-      event: 'article_tag_has_been_removed',
+      event: DB_NOTICE_TYPE.article_tag_has_been_removed,
       recipientId: article.authorId,
       actorId: viewer.id,
       entities: [
