+++ src/mutations/article/updateArticleState.ts
@@ -1,4 +1,4 @@
-import { ARTICLE_STATE } from 'common/enums'
+import { ARTICLE_STATE, OFFICIAL_NOTICE_EXTEND_TYPE } from 'common/enums'
 import { fromGlobalId } from 'common/utils'
 import { MutationToUpdateArticleStateResolver } from 'definitions'
 
@@ -26,7 +26,7 @@ const resolver: MutationToUpdateArticleStateResolver = async (
   if (state === ARTICLE_STATE.banned) {
     const user = await userService.dataloader.load(article.authorId)
     notificationService.trigger({
-      event: 'article_banned',
+      event: OFFICIAL_NOTICE_EXTEND_TYPE.article_banned,
       entities: [{ type: 'target', entityTable: 'article', entity: article }],
       recipientId: user.id,
     })
